+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{
+    "files.associations": {
+        "*.ejs": "html",
+        "*.html": "html",
+        "vector": "cpp",
+        "ostream": "cpp"
+    }
+}
